requirements.txt
boto3==1.34.113
librosa==0.10.1
mage-ai==0.9./&
mlflow==2.13.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
ruff==0.4.5
scikit-learn==1.7.1
